+++ DAO/UDTCourseDef.cs
@@ -57,7 +57,7 @@ namespace SHSchool.Retake.DAO
         /// 學分數
         ///</summary>
         [Field(Field = "credit", Indexed = false)]
-        public int Credit { get; set; }
+        public decimal Credit { get; set; }
 
         ///<summary>
         /// 授課教師名稱(存 ID)
